refactor: update Dockerfile to use Bun for building and managing dependencies; change Docker Compose target to 'release'; adjust import paths in auth and user modules; remove unused validation utility
