+++ src/modules/auth/dto/auth.dto.ts
@@ -1,3 +1,3 @@
-import { CreateUserDto } from 'src/modules/user/dto/create-user.dto';
+import { CreateUserDto } from '../../user/dto/create-user.dto';
 
 export type UserAuth = Omit<CreateUserDto, 'password'> & { id: number };
