+++ src/modules/logos/logos.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/modules/prisma/prisma.module';
+import { PrismaModule } from '../prisma/prisma.module';
 import { LogosController } from './logos.controller';
 import { LogosService } from './logos.service';
